+++ Makefile
@@ -4897,7 +4897,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/impl/codegen/client_callback.h \
     include/grpcpp/impl/codegen/client_callback_impl.h \
     include/grpcpp/impl/codegen/client_context.h \
-    include/grpcpp/impl/codegen/client_context_impl.h \
     include/grpcpp/impl/codegen/client_interceptor.h \
     include/grpcpp/impl/codegen/client_unary_call.h \
     include/grpcpp/impl/codegen/completion_queue.h \
@@ -5581,7 +5580,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/impl/codegen/client_callback.h \
     include/grpcpp/impl/codegen/client_callback_impl.h \
     include/grpcpp/impl/codegen/client_context.h \
-    include/grpcpp/impl/codegen/client_context_impl.h \
     include/grpcpp/impl/codegen/client_interceptor.h \
     include/grpcpp/impl/codegen/client_unary_call.h \
     include/grpcpp/impl/codegen/completion_queue.h \
