+++ gRPC-C++.podspec
@@ -106,7 +106,6 @@ Pod::Spec.new do |s|
                       'include/grpcpp/impl/codegen/client_callback.h',
                       'include/grpcpp/impl/codegen/client_callback_impl.h',
                       'include/grpcpp/impl/codegen/client_context.h',
-                      'include/grpcpp/impl/codegen/client_context_impl.h',
                       'include/grpcpp/impl/codegen/client_interceptor.h',
                       'include/grpcpp/impl/codegen/client_unary_call.h',
                       'include/grpcpp/impl/codegen/completion_queue.h',
