+++ include/grpcpp/channel.h
@@ -86,7 +86,7 @@ class Channel final : public ::grpc::ChannelInterface,
               interceptor_creators);
 
   ::grpc::internal::Call CreateCall(const ::grpc::internal::RpcMethod& method,
-                                    ::grpc_impl::ClientContext* context,
+                                    ::grpc::ClientContext* context,
                                     ::grpc::CompletionQueue* cq) override;
   void PerformOpsOnCall(::grpc::internal::CallOpSetInterface* ops,
                         ::grpc::internal::Call* call) override;
@@ -101,9 +101,8 @@ class Channel final : public ::grpc::ChannelInterface,
   ::grpc::CompletionQueue* CallbackCQ() override;
 
   ::grpc::internal::Call CreateCallInternal(
-      const ::grpc::internal::RpcMethod& method,
-      ::grpc_impl::ClientContext* context, ::grpc::CompletionQueue* cq,
-      size_t interceptor_pos) override;
+      const ::grpc::internal::RpcMethod& method, ::grpc::ClientContext* context,
+      ::grpc::CompletionQueue* cq, size_t interceptor_pos) override;
 
   const std::string host_;
   grpc_channel* const c_channel_;  // owned
