+++ include/grpcpp/impl/codegen/client_callback_impl.h
@@ -30,13 +30,13 @@
 
 namespace grpc {
 class Channel;
+class ClientContext;
 namespace internal {
 class RpcMethod;
 }  // namespace internal
 }  // namespace grpc
 
 namespace grpc_impl {
-class ClientContext;
 
 namespace internal {
 
@@ -45,7 +45,7 @@ namespace internal {
 template <class InputMessage, class OutputMessage>
 void CallbackUnaryCall(::grpc::ChannelInterface* channel,
                        const ::grpc::internal::RpcMethod& method,
-                       ::grpc_impl::ClientContext* context,
+                       ::grpc::ClientContext* context,
                        const InputMessage* request, OutputMessage* result,
                        std::function<void(::grpc::Status)> on_completion) {
   CallbackUnaryCallImpl<InputMessage, OutputMessage> x(
@@ -57,7 +57,7 @@ class CallbackUnaryCallImpl {
  public:
   CallbackUnaryCallImpl(::grpc::ChannelInterface* channel,
                         const ::grpc::internal::RpcMethod& method,
-                        ::grpc_impl::ClientContext* context,
+                        ::grpc::ClientContext* context,
                         const InputMessage* request, OutputMessage* result,
                         std::function<void(::grpc::Status)> on_completion) {
     ::grpc::CompletionQueue* cq = channel->CallbackCQ();
@@ -574,7 +574,7 @@ class ClientCallbackReaderWriterImpl
   friend class ClientCallbackReaderWriterFactory<Request, Response>;
 
   ClientCallbackReaderWriterImpl(grpc::internal::Call call,
-                                 ::grpc_impl::ClientContext* context,
+                                 ::grpc::ClientContext* context,
                                  ClientBidiReactor<Request, Response>* reactor)
       : context_(context),
         call_(call),
@@ -641,7 +641,7 @@ class ClientCallbackReaderWriterImpl
     }
   }
 
-  ::grpc_impl::ClientContext* const context_;
+  ::grpc::ClientContext* const context_;
   grpc::internal::Call call_;
   ClientBidiReactor<Request, Response>* const reactor_;
 
@@ -690,7 +690,7 @@ class ClientCallbackReaderWriterFactory {
  public:
   static void Create(::grpc::ChannelInterface* channel,
                      const ::grpc::internal::RpcMethod& method,
-                     ::grpc_impl::ClientContext* context,
+                     ::grpc::ClientContext* context,
                      ClientBidiReactor<Request, Response>* reactor) {
     grpc::internal::Call call =
         channel->CreateCall(method, context, channel->CallbackCQ());
@@ -785,8 +785,7 @@ class ClientCallbackReaderImpl : public ClientCallbackReader<Response> {
 
   template <class Request>
   ClientCallbackReaderImpl(::grpc::internal::Call call,
-                           ::grpc_impl::ClientContext* context,
-                           Request* request,
+                           ::grpc::ClientContext* context, Request* request,
                            ClientReadReactor<Response>* reactor)
       : context_(context), call_(call), reactor_(reactor) {
     this->BindReactor(reactor);
@@ -812,7 +811,7 @@ class ClientCallbackReaderImpl : public ClientCallbackReader<Response> {
     }
   }
 
-  ::grpc_impl::ClientContext* const context_;
+  ::grpc::ClientContext* const context_;
   grpc::internal::Call call_;
   ClientReadReactor<Response>* const reactor_;
 
@@ -848,8 +847,7 @@ class ClientCallbackReaderFactory {
   template <class Request>
   static void Create(::grpc::ChannelInterface* channel,
                      const ::grpc::internal::RpcMethod& method,
-                     ::grpc_impl::ClientContext* context,
-                     const Request* request,
+                     ::grpc::ClientContext* context, const Request* request,
                      ClientReadReactor<Response>* reactor) {
     grpc::internal::Call call =
         channel->CreateCall(method, context, channel->CallbackCQ());
@@ -970,8 +968,7 @@ class ClientCallbackWriterImpl : public ClientCallbackWriter<Request> {
 
   template <class Response>
   ClientCallbackWriterImpl(::grpc::internal::Call call,
-                           ::grpc_impl::ClientContext* context,
-                           Response* response,
+                           ::grpc::ClientContext* context, Response* response,
                            ClientWriteReactor<Request>* reactor)
       : context_(context),
         call_(call),
@@ -1027,7 +1024,7 @@ class ClientCallbackWriterImpl : public ClientCallbackWriter<Request> {
     }
   }
 
-  ::grpc_impl::ClientContext* const context_;
+  ::grpc::ClientContext* const context_;
   grpc::internal::Call call_;
   ClientWriteReactor<Request>* const reactor_;
 
@@ -1074,7 +1071,7 @@ class ClientCallbackWriterFactory {
   template <class Response>
   static void Create(::grpc::ChannelInterface* channel,
                      const ::grpc::internal::RpcMethod& method,
-                     ::grpc_impl::ClientContext* context, Response* response,
+                     ::grpc::ClientContext* context, Response* response,
                      ClientWriteReactor<Request>* reactor) {
     grpc::internal::Call call =
         channel->CreateCall(method, context, channel->CallbackCQ());
@@ -1130,7 +1127,7 @@ class ClientCallbackUnaryImpl final : public ClientCallbackUnary {
 
   template <class Request, class Response>
   ClientCallbackUnaryImpl(::grpc::internal::Call call,
-                          ::grpc_impl::ClientContext* context, Request* request,
+                          ::grpc::ClientContext* context, Request* request,
                           Response* response, ClientUnaryReactor* reactor)
       : context_(context), call_(call), reactor_(reactor) {
     this->BindReactor(reactor);
@@ -1156,7 +1153,7 @@ class ClientCallbackUnaryImpl final : public ClientCallbackUnary {
     }
   }
 
-  ::grpc_impl::ClientContext* const context_;
+  ::grpc::ClientContext* const context_;
   grpc::internal::Call call_;
   ClientUnaryReactor* const reactor_;
 
@@ -1182,9 +1179,8 @@ class ClientCallbackUnaryFactory {
   template <class Request, class Response>
   static void Create(::grpc::ChannelInterface* channel,
                      const ::grpc::internal::RpcMethod& method,
-                     ::grpc_impl::ClientContext* context,
-                     const Request* request, Response* response,
-                     ClientUnaryReactor* reactor) {
+                     ::grpc::ClientContext* context, const Request* request,
+                     Response* response, ClientUnaryReactor* reactor) {
     grpc::internal::Call call =
         channel->CreateCall(method, context, channel->CallbackCQ());
