+++ include/grpcpp/impl/codegen/client_interceptor.h
@@ -26,14 +26,10 @@
 #include <grpcpp/impl/codegen/rpc_method.h>
 #include <grpcpp/impl/codegen/string_ref.h>
 
-namespace grpc_impl {
-
-class ClientContext;
-}  // namespace grpc_impl
-
 namespace grpc {
 
 class Channel;
+class ClientContext;
 
 namespace internal {
 class InterceptorBatchMethodsImpl;
@@ -96,7 +92,7 @@ class ClientRpcInfo {
 
   /// Return a pointer to the underlying ClientContext structure associated
   /// with the RPC to support features that apply to it
-  grpc_impl::ClientContext* client_context() { return ctx_; }
+  grpc::ClientContext* client_context() { return ctx_; }
 
   /// Return the type of the RPC (unary or a streaming flavor)
   Type type() const { return type_; }
@@ -119,9 +115,8 @@ class ClientRpcInfo {
   ClientRpcInfo() = default;
 
   // Constructor will only be called from ClientContext
-  ClientRpcInfo(grpc_impl::ClientContext* ctx,
-                internal::RpcMethod::RpcType type, const char* method,
-                grpc::ChannelInterface* channel)
+  ClientRpcInfo(grpc::ClientContext* ctx, internal::RpcMethod::RpcType type,
+                const char* method, grpc::ChannelInterface* channel)
       : ctx_(ctx),
         type_(static_cast<Type>(type)),
         method_(method),
@@ -163,7 +158,7 @@ class ClientRpcInfo {
     }
   }
 
-  grpc_impl::ClientContext* ctx_ = nullptr;
+  grpc::ClientContext* ctx_ = nullptr;
   // TODO(yashykt): make type_ const once move-assignment is deleted
   Type type_{Type::UNKNOWN};
   const char* method_ = nullptr;
@@ -173,7 +168,7 @@ class ClientRpcInfo {
   size_t hijacked_interceptor_ = 0;
 
   friend class internal::InterceptorBatchMethodsImpl;
-  friend class grpc_impl::ClientContext;
+  friend class grpc::ClientContext;
 };
 
 // PLEASE DO NOT USE THIS. ALWAYS PREFER PER CHANNEL INTERCEPTORS OVER A GLOBAL
