+++ include/grpcpp/impl/codegen/intercepted_channel.h
@@ -46,7 +46,7 @@ class InterceptedChannel : public ChannelInterface {
   InterceptedChannel(ChannelInterface* channel, size_t pos)
       : channel_(channel), interceptor_pos_(pos) {}
 
-  Call CreateCall(const RpcMethod& method, ::grpc_impl::ClientContext* context,
+  Call CreateCall(const RpcMethod& method, ::grpc::ClientContext* context,
                   ::grpc::CompletionQueue* cq) override {
     return channel_->CreateCallInternal(method, context, cq, interceptor_pos_);
   }
